pyproject.toml
# full spec see here: https://json.schemastore.org/pyproject.json

[project]
name = "sortipy"
dynamic = ["version"]
authors = [{ name = "Simon Fischer", email = "sortipy@simon-fischer.info" }]
description = "Spotify stuff"
license = { file = "LICENSE" }
requires-python = ">= 3.14"
readme = "README.md"
dependencies = [
    "spotipy",
    "httpx",
    "httpx-retries",
    "sqlalchemy",
    "aiolimiter",
    "hishel[async]",
    "pydantic>=2.8,<3",
    "alembic"
]
scripts.sortipy = "sortipy.ui.cli:main"


[dependency-groups]
dev = [
  "coverage[toml]",
  "poethepoet>=0.40.0",
  "pytest>=9.0.0",
  "pytest-asyncio>=1.3.0",
  "pytest-cov",
  "python-dotenv>=1.2.1",
  "ruff>=0.14.13",
  "pyright",
]


[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"


[tool.hatch]
version.source = "vcs"


[tool.poe.tasks]
format = "ruff format"
format_check = "ruff format --check"
lint_fix = "ruff check --fix"
lint = "ruff check"
typecheck = "pyright"
test = "pytest --exitfirst"
test_cov = "pytest --cov --cov-report=term --cov-report=html"
fix = ["format", "lint_fix"]
check = ["format_check", "lint", "typecheck", "test_cov"]


[tool.pytest]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
strict = true
addopts = ["--log-level=DEBUG", "--color=yes", "--failed-first"]
testpaths = ["tests"]
pythonpath = ["src"]
markers = ["integration: test is an integration test"]


[tool.coverage]
run.source = ["src"]
run.branch = true
run.parallel = true
report.exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:"
]
report.skip_covered = true
report.precision = 1
report.fail_under = 0
html.directory = "htmlcov"


[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py314"
lint.select = ["ALL"]
lint.extend-safe-fixes = ["TC001", "TC002", "TC003", "TC004"]
lint.future-annotations = true
lint.ignore = [
  "CPY",    # missing copyright notice
  "G004",   # f-strings in log messages
  "FIX",    # presence of todo
  "TD",     # formatting of todo
  "EM",     # error message formatting
  "DOC",    # pydoclint
  "D",      # pydocstyle
  "TRY003", # long exception message
  # see https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules:
  "W191",
  "E111",
  "E114",
  "E117",
  "D206",
  "D300",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "COM812",
  "COM819",
]
lint.per-file-ignores = { "tests/*" = [
  "INP001",  # missing _init_.py
  "ARG002",  # unused method argument
  "S101",    # asserts in tests
  "ANN001",  # missing arg type in tests
  "ANN201",  # missing return type in tests
  "ANN202",  # missing return type for private funcs in tests
  "SLF001",  # private access
  "PLR2004", # magic values
], "src/sortipy/adapters/sqlalchemy/migrations/versions/*.py" = [
    "I001", # generated migration import ordering
] }
lint.flake8-type-checking.runtime-evaluated-base-classes = [
  "pydantic.BaseModel"
]
lint.pylint.max-args = 7


[tool.pyright] # see https://github.com/microsoft/pyright/blob/main/docs/configuration.md
include = ["src", "tests"]
ignore = ["**/__pycache__", ".venv"]
reportPrivateUsage = "none"  # handled by ruff
typeCheckingMode = "strict"
pythonVersion = "3.14"
pythonPlatform = "Linux"
venvPath = "."
venv = ".venv"


[tool.alembic]
script_location = "%(here)s/src/sortipy/adapters/sqlalchemy/migrations"
prepend_sys_path = ["%(here)s/src"]   # statt "."
